+++ bin/micrograd_basic.ml
@@ -13,9 +13,9 @@ let _suspended () =
   CDSL.with_debug := true;
   CDSL.keep_files_in_run_directory := true;
   Random.init 0;
-  let%nn_op c = "a" [ -4 ] + "b" [ 2 ] in
-  (* let%nn_op c = c + c + 1 in
-     let%nn_op c = c + 1 + c + ~-a in *)
+  let%op c = "a" [ -4 ] + "b" [ 2 ] in
+  (* let%op c = c + c + 1 in
+     let%op c = c + 1 + c + ~-a in *)
   (* Tensor.set_fully_on_host g;
      Tensor.set_fully_on_host a;
      Tensor.set_fully_on_host b; *)
@@ -33,16 +33,16 @@ let () =
   CDSL.with_debug := true;
   CDSL.keep_files_in_run_directory := true;
   Random.init 0;
-  let%nn_op c = "a" [ -4 ] + "b" [ 2 ] in
-  let%nn_op d = (a *. b) + (b **. 3) in
-  let%nn_op c = c + c + 1 in
-  let%nn_op c = c + 1 + c + ~-a in
-  let%nn_op d = d + (d *. 2) + !/(b + a) in
-  let%nn_op d = d + (3 *. d) + !/(b - a) in
-  let%nn_op e = c - d in
-  let%nn_op f = e *. e in
-  let%nn_op g = f /. 2 in
-  let%nn_op g = g + (10. /. f) in
+  let%op c = "a" [ -4 ] + "b" [ 2 ] in
+  let%op d = (a *. b) + (b **. 3) in
+  let%op c = c + c + 1 in
+  let%op c = c + 1 + c + ~-a in
+  let%op d = d + (d *. 2) + !/(b + a) in
+  let%op d = d + (3 *. d) + !/(b - a) in
+  let%op e = c - d in
+  let%op f = e *. e in
+  let%op g = f /. 2 in
+  let%op g = g + (10. /. f) in
   (* *)
   Tensor.set_fully_on_host g;
   Tensor.set_fully_on_host a;