+++ bin/micrograd_demo.ml
@@ -46,19 +46,19 @@ let () =
       ~o:[ 1 ]
       moons_classes
   in
-  let%nn_op mlp x = "b3" 1 + ("w3" * !/("b2" hid_dim + ("w2" * !/("b1" hid_dim + ("w1" * x))))) in
+  let%op mlp x = "b3" 1 + ("w3" * !/("b2" hid_dim + ("w2" * !/("b1" hid_dim + ("w1" * x))))) in
   let session_step = NTDSL.O.(NTDSL.counter !..1) in
-  let%nn_op minus_lr = -0.1 *. (!..steps - session_step) /. !..steps in
+  let%op minus_lr = -0.1 *. (!..steps - session_step) /. !..steps in
   (* minus_learning_rate := Some minus_lr; *)
-  let%nn_op moons_input = moons_flat @.| session_step in
-  let%nn_op moons_class = moons_classes @.| session_step in
+  let%op moons_input = moons_flat @.| session_step in
+  let%op moons_class = moons_classes @.| session_step in
   let losses = ref [] in
   let log_losses = ref [] in
   let learning_rates = ref [] in
-  let%nn_op margin_loss = !/(1 - (moons_class *. mlp moons_input)) in
-  let%nn_op ssq w = (w **. 2) ++ "...|...->... => 0" in
+  let%op margin_loss = !/(1 - (moons_class *. mlp moons_input)) in
+  let%op ssq w = (w **. 2) ++ "...|...->... => 0" in
   let reg_loss = List.map ~f:ssq [ w1; w2; w3; b1; b2; b3 ] |> List.reduce_exn ~f:TDSL.O.( + ) in
-  let%nn_op total_loss = ((margin_loss ++ "...|... => 0") /. !..batch) + (0.0001 *. reg_loss) in
+  let%op total_loss = ((margin_loss ++ "...|... => 0") /. !..batch) + (0.0001 *. reg_loss) in
   (* SDSL.everything_on_host_or_inlined (); *)
   for step = 1 to steps do
     (* refresh_session (); *)
@@ -76,7 +76,7 @@ let () =
   let classes = Tensor.value_1d_points ~xdim:0 moons_classes in
   let points1, points2 = Array.partitioni_tf points ~f:Float.(fun i _ -> classes.(i) > 0.) in
   SDSL.close_session ();
-  let%nn_op point = [ 0; 0 ] in
+  let%op point = [ 0; 0 ] in
   let mlp_result = mlp point in
   SDSL.refresh_session ~with_backprop:false ();
   let callback (x, y) =