+++ bin/zero2hero_1of7.ml
@@ -10,7 +10,7 @@ module NTDSL = Operation.NTDSL
 let _suspended () =
   (* SDSL.drop_all_sessions (); *)
   Random.init 0;
-  let%nn_op v = ("w" [ (-3, 1) ] * "x" [ 2; 0 ]) + "b" [ 6.7 ] in
+  let%op v = ("w" [ (-3, 1) ] * "x" [ 2; 0 ]) + "b" [ 6.7 ] in
   (* refresh_session (); *)
   Stdio.printf "\n%!";
   Tensor.print_tree ~with_id:true ~with_grad:true ~depth:9 v;
@@ -25,8 +25,8 @@ let _suspended () =
   Random.init 0;
   CDSL.enable_all_debugs ();
   CDSL.virtualize_settings.enable_device_only <- false;
-  let%nn_op f x = (3 *. (x **. 2)) - (4 *. x) + 5 in
-  let%nn_op f5 = f 5 in
+  let%op f x = (3 *. (x **. 2)) - (4 *. x) + 5 in
+  let%op f5 = f 5 in
   (* refresh_session (); *)
   Stdio.printf "\n%!";
   Tensor.print_tree ~with_grad:false ~depth:9 f5;
@@ -36,7 +36,7 @@ let _suspended () =
   let open Tensor.O in
   (* SDSL.drop_all_sessions (); *)
   Random.init 0;
-  let%nn_op f x = (3 *. (x **. 2)) - (4 *. x) + 5 in
+  let%op f x = (3 *. (x **. 2)) - (4 *. x) + 5 in
   let size = 100 in
   let xs = Array.init size ~f:Float.(fun i -> (of_int i / 10.) - 5.) in
   let x_flat =
@@ -47,8 +47,8 @@ let _suspended () =
       ~init_op:(Constant_fill xs) ()
   in
   let session_step = NTDSL.O.(NTDSL.counter !..1) in
-  let%nn_op x = x_flat @.| session_step in
-  let%nn_op fx = f x in
+  let%op x = x_flat @.| session_step in
+  let%op fx = f x in
   Stdio.print_endline "\n";
   Tensor.print_tree ~with_id:true ~with_value:false ~with_grad:false ~depth:9 fx;
   Stdio.print_endline "\n";
@@ -87,14 +87,14 @@ let _suspended () =
      CDSL.keep_files_in_run_directory := true; *)
   Random.init 0;
   Stdio.print_endline "\nFirst refresh:";
-  let%nn_op f = (3 *. ("x" [ 5 ] **. 2)) - (4 *. x) + 5 in
+  let%op f = (3 *. ("x" [ 5 ] **. 2)) - (4 *. x) + 5 in
   (* refresh_session (); *)
   Tensor.print_tree ~with_grad:true ~depth:9 f;
   let xs = Array.init 10 ~f:Float.(fun i -> of_int i - 5.) in
   let ys =
     Array.map xs ~f:(fun v ->
         (* This is inefficient because it compiles the argument update inside the loop. *)
-        SDSL.compile_routine [%nn_cd x =: !.v] ~name:"assign_x" ();
+        SDSL.compile_routine [%cd x =: !.v] ~name:"assign_x" ();
         (* refresh_session (); *)
         f.@[0])
   in
@@ -110,9 +110,9 @@ let () =
   CDSL.with_debug := true;
   CDSL.keep_files_in_run_directory := true;
   Random.init 0;
-  let%nn_op e = "a" [ 2 ] *. "b" [ -3 ] in
-  let%nn_op d = e + "c" [ 10 ] in
-  let%nn_op l = d *. "f" [ -2 ] in
+  let%op e = "a" [ 2 ] *. "b" [ -3 ] in
+  let%op d = e + "c" [ 10 ] in
+  let%op l = d *. "f" [ -2 ] in
   SDSL.minus_learning_rate := Some (TDSL.init_const ~l:"minus_lr" ~o:[ 1 ] [| 0.1 |]);
   (* everything_fully_on_host (); *)
   SDSL.refresh_session ~update_params:false ();