+++ lib/ppx_op.ml
@@ -1,7 +1,7 @@
 open Base
 open Ppxlib
 open Ppx_arrayjit.Ppx_helper
-open Ppx_nn_shared
+open Ppx_shared
 
 let ndarray_op ?desc_label ?axis_labels ?label expr =
   let loc = expr.pexp_loc in