+++ CHANGES.md
@@ -9,6 +9,7 @@ Major rewrite.
 - No more "form vs. non-form" formulas / tensors.
   - Formula/tensor roots are split into forward roots and backprop roots.
 - No more `%nn_rs`, `%nn_dt` syntaxes and `Synthetic` fetch primitive.
+- Renamed `%nn_op` to `%op` and `%nn_cd` to `%cd`.
 
 ## [0.2.1] -- 2023-07-19